:hammer: Add business rules

+++ back-end/src/services/taskService.js
@@ -1,12 +1,9 @@
 const { ObjectId } = require('mongodb');
 const taskModel = require('../models/taskModel');
+const errorDefault = require('../error/errorDefault');
 
 const getAllTasks = async () => {
   const tasks = await taskModel.getAllTasks();
-
-  // const tasksNameMap = tasks.map(({ taskName }) => ({ taskName: taskName.toLowerCase() }));
-  // const tasksNameSort = tasksNameMap.sort();
-  // console.log(tasksNameSort);
 
   return tasks;
 };
@@ -14,17 +11,27 @@ const getAllTasks = async () => {
 const addTask = async (taskName) => {
   const taskId = await taskModel.addTask(taskName);
 
+  if (!taskName) throw (errorDefault(400, 'Task Name is required'));
+
   return taskId;
 };
 
 const updateTask = async (id, taskName) => {
+  if (!ObjectId.isValid(id)) throw (errorDefault(400, 'This Id is invalid'));
+
   const task = await taskModel.updateTask(new ObjectId(id), taskName);
+
+  if (!task) throw (errorDefault(404, 'Task not found'));
 
   return task;
 };
 
 const deleteTask = async (id) => {
+  if (!ObjectId.isValid(id)) throw (errorDefault(400, 'This Id is invalid'));
+
   const task = await taskModel.deleteTask(new ObjectId(id));
+
+  if (!task) throw (errorDefault(404, 'Task not found'));
 
   return task;
 };
